Suppress messages when grid has no owner
if MySession.Static.Players.TryGetSteamId(PB.OwnerId) doesn't work, the message is sent to ID:0, which is everyone on the server. Pretty annoying when there's a bunch of reavers flying around. Let's suppress the message if there's no owner, perhaps? Thoughts appreciated. I figure this could be a checkbox in the UI as well, "[] Suppress messages for ownerless grids?" I'm unfortunately not well versed enough to try to do a pull request for that.

Sidenote, not sure if MySession.Static.Players.TryGetSteamId(PB.OwnerId) is a double, decimal or what. I assumed double, could be wrong.